[JENKINS-53785] Convert DockerComputerConnector$1 to non anymous class
Seen through Sentry:

```
Attempt to (de-)serialize anonymous class io.jenkins.docker.connector.DockerComputerConnector$1 in file:/evergreen/data/jenkins/var/plugins/docker-plugin/WEB-INF/lib/docker-plugin.jar; see: https://jenkins.io/redirect/serialization-of-anonymous-classes/

```
